+++ comfyui_manager/common/cm_global.py
