+++ comfyui_manager/common/enums.py
